# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Spanish Tutor Booking Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /files/homework/{homeworkId}/{which}:
    get:
      operationId: getHomeworkFile
      tags: [uploads]
      summary: Download/view a homework file (assigned, submission, or review) via a private proxy
      parameters:
        - name: homeworkId
          in: path
          required: true
          schema:
            type: number
        - name: which
          in: path
          required: true
          schema:
            type: string
            enum: [assigned, submission, review]
      responses:
        "200":
          description: File contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "401":
          description: Unauthorized
        "404":
          description: Not found
...